--- source/databricks/calculation_engine/package/contracts/dataframes/columns/column.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-from pyspark.sql.types import DataType
-
-
-@dataclass
-class Column:
-    name: str
-    data_type: DataType
